Add files via upload
A home assistance like Alexa, but gesture/human behavior controlled. We will use only a raspberry pi and a camera to predict what the person wants to do and then execute that prediction. This device will have the ability to learn over time and the person will be able to teach it new skills.